backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ml-dtypes>=0.4.0",
    "nemo-toolkit[asr]>=2.3.0",
    "numpy<2.0",
    "onnx>=1.16.1",
    "onnxruntime>=1.17.0",
    "PySide6>=6.8.0",
    "pynput>=1.8.1",
    "sounddevice>=0.5.1",
    "torch>=2.0.0",
    "faster-whisper>=1.2.1",
]
